使用迭代方法实现Geometry的渲染方法的遍历，解除了渲染逻辑的杂糅，为Geometry定义drawAll()方法进行遍历，在遍历中为每个节点调用各自的draw()方法，draw()方法不考虑树状结构，只封装渲染当前节点的逻辑

+++ project_test/project/interaction.cpp
@@ -48,7 +48,7 @@ void mouse_callback(GLFWwindow* window, double xpos, double ypos) {
 			//camera->move(-(float)dx, (float)dy, 0.0f);
 			//camera->lookAt(glm::vec3(0.0f, 0.0f, 0.0f));
 			// 以(0.0,1.0,0.0)为轴，旋转相机，包括相机的front向量一同旋转
-			camera->rotateByAxis(dx, dy);
+			camera->rotateByAxis((float)dx, (float)dy);
 		}
 	}
 	if (status.leftMouseButtonPressed == true) {
